Color background, fix score board, & fix typos
This added a simple solid background color. It also fixed the score
board since it the names and the actual score were buggy due to the JS
code. Then I fixed some typos in the JS, I changed "Human" to "You"
since it makes a bit more sense to refer to the user in this way.