Move the player that one next to score board
Orginally the player that won would be displayed at the bottom. This
was weird for phones and especially laptops screens. So I moved that
to be next to the score board. I also made it so who ever wins will
either show on the left to indicate the user won the round or right
for the computer. By left and right I mean in relation to the score
board.